Fix: Module import issues
