+++ n_to_n_matching/src/n_to_n_matching/__main__.py
@@ -1,4 +1,11 @@
-from n_to_n_matching.mather import test_1
+from n_to_n_matching.test_main import test_1
+
+import sys
+
+def main():
+    # Check Python environment
+    print("Python sys.path: {}".format(sys.path))
+    test_1()
 
 if __name__ == "__main__":
-    test_1()
+    main()
