+++ pyproject.toml
@@ -17,6 +17,7 @@ classifiers = [
 keywords = ["many-to-many matching"]
 dependencies = [
     "matching",
+    "PyYAML>=6.0.1"
 ]
 requires-python = ">=3.9"
 
@@ -30,4 +31,4 @@ Homepage = "https://github.com/kinu-garage/nton_matching"
 gjls_match = "n_to_n_matching.__main__:main"
 
 [tool.setuptools]
-package-dir = {"" = "src"}
+package-dir = {"" = "n_to_n_matching/src"}
